bpf: tests: condense LB wildcard-drop test
Run the IPv4 and IPv6 test from a single file.
